requirements.txt
browser-use
openai
python-dotenv
PyPDF2
langchain-openai
